payment api

+++ Backend/src/main/java/com/group07/PetHealthCare/dto/request/AppointmentRequest.java
@@ -14,7 +14,7 @@ import java.util.Set;
 public class AppointmentRequest {
     private String status;
     private LocalDate appointmentDate;
-    private BigDecimal deposit;
+    private Double deposit;
     private String petId;
     private Set<String> serviceId;
     private String veterinarianId;
